Move inner functions from ScalarTypesProvider to a separate utility file
Phpstan doesn't understand inner functions
