+++ composer.json
@@ -26,6 +26,7 @@
             "./src/Type/directives.php",
             "./src/Type/introspection.php",
             "./src/Type/scalars.php",
+            "./src/Util/coercers.php",
             "./src/Util/helpers.php",
             "./src/Util/invariant.php",
             "./src/Util/valueFromAST.php",
